feat: fix database type

+++ src/main/java/org/management/core/application/common/param/vo/UserVO.java
@@ -13,6 +13,6 @@ import java.io.Serializable;
 public class UserVO implements Serializable {
     
     private Integer userId;
-    private Byte userType;
+    private Integer userType;
     private String token;
 }
